Feature/AudioInspector - Routing for direct out and patches now working.
Took 6 hours 18 minutes

+++ src/Shared/MediaEngine.cs
@@ -72,7 +72,7 @@ public partial class MediaEngine : Node
                 var primaryAudio = audioTracks.First();
 
                 // Number of channels
-                metadata.Add("Channels", primaryAudio.Data.Audio.Channels);
+                metadata.Add("Channels", (int)primaryAudio.Data.Audio.Channels);
 
                 // Sample rate (Hz)
                 if (primaryAudio.Data.Audio.Rate > 0)
